+++ src/common/gitProvider/azureDevops.ts
@@ -30,9 +30,8 @@ export class AzureDevopsProvider extends GitProviderRoot {
   // Returns current job URL
   public async getCurrentJobUrl(): Promise<string> {
     if (process.env.SYSTEM_COLLECTIONURI && process.env.SYSTEM_TEAMPROJECT && process.env.BUILD_BUILDID) {
-      const jobUrl = `${process.env.SYSTEM_COLLECTIONURI}${encodeURIComponent(process.env.SYSTEM_TEAMPROJECT)}/_build/results?buildId=${
-        process.env.BUILD_BUILDID
-      }`;
+      const jobUrl = `${process.env.SYSTEM_COLLECTIONURI}${encodeURIComponent(process.env.SYSTEM_TEAMPROJECT)}/_build/results?buildId=${process.env.BUILD_BUILDID
+        }`;
       return jobUrl;
     }
     uxLog(
@@ -61,25 +60,7 @@ export class AzureDevopsProvider extends GitProviderRoot {
     uxLog(
       this,
       c.yellow(`[Azure DevOps] You need the following variables to be defined in azure devops pipeline step:
-      SFDX_DEPLOY_WAIT_MINUTES: 150
-      CI_COMMIT_REF_NAME: $(BRANCH_NAME)
-      CONFIG_BRANCH: $(BRANCH_NAME)
-      ORG_ALIAS: $(BRANCH_NAME)
-      SLACK_TOKEN: $(SLACK_TOKEN)
-      SLACK_CHANNEL_ID: $(SLACK_CHANNEL_ID)
-      MS_TEAMS_WEBHOOK_URL: $(MS_TEAMS_WEBHOOK_URL)
-      CI: "true"
-      SYSTEM_ACCESSTOKEN: $(System.AccessToken)
-      CI_SFDX_HARDIS_AZURE_TOKEN: $(System.AccessToken)
-      SYSTEM_COLLECTIONURI: $(System.CollectionUri)
-      SYSTEM_TEAMPROJECT: $(System.TeamProject)
-      SYSTEM_JOB_DISPLAY_NAME: $(System.JobDisplayName)
-      SYSTEM_JOB_ID: $(System.JobId)
-      SYSTEM_PULLREQUEST_PULLREQUESTID: $(System.PullRequest.PullRequestId)
-      BUILD_REPOSITORY_ID: $(Build.Repository.ID)
-      BUILD_REPOSITORYNAME: $(Build.Repository.Name)
-      BUILD_SOURCEBRANCHNAME: $(Build.SourceBranchName)
-      BUILD_BUILD_ID: $(Build.BuildId)
+${this.getPipelineVariablesConfig()}
 `),
     );
     return null;
@@ -170,14 +151,8 @@ export class AzureDevopsProvider extends GitProviderRoot {
       uxLog(this, c.grey("[Azure integration] No project and pull request, so no note thread..."));
       uxLog(
         this,
-        c.yellow(`Following variables must be defined when available:
-- BUILD_REPOSITORY_ID
-- BUILD_BUILD_ID
-- SYSTEM_JOB_ID
-- SYSTEM_PULLREQUEST_PULLREQUESTID
-- SYSTEM_JOB_DISPLAY_NAME
-- SYSTEM_COLLECTIONURI
-- SYSTEM_TEAMPROJECT
+        c.yellow(`Following variables should be defined when available:
+${this.getPipelineVariablesConfig()}
       `),
       );
       return { posted: false, providerResult: { info: "No related pull request" } };
@@ -269,6 +244,30 @@ _Provided by [sfdx-hardis](https://sfdx-hardis.cloudity.com) from job [${azureJo
     const prInfo: any = Object.assign({}, prData);
     prInfo.sourceBranch = (prData.sourceRefName || "").replace("refs/heads/", "");
     prInfo.targetBranch = (prData.targetRefName || "").replace("refs/heads/", "");
+    prInfo.web_url = `${process.env.SYSTEM_COLLECTIONURI}${encodeURIComponent(process.env.SYSTEM_TEAMPROJECT)}/_git/${encodeURIComponent(process.env.BUILD_REPOSITORYNAME)}/pullrequest/${prData.pullRequestId}`;
     return prInfo;
+  }
+
+  private getPipelineVariablesConfig() {
+    return `
+    SFDX_DEPLOY_WAIT_MINUTES: 150
+    CI_COMMIT_REF_NAME: $(BRANCH_NAME)
+    CONFIG_BRANCH: $(BRANCH_NAME)
+    ORG_ALIAS: $(BRANCH_NAME)
+    SLACK_TOKEN: $(SLACK_TOKEN)
+    SLACK_CHANNEL_ID: $(SLACK_CHANNEL_ID)
+    MS_TEAMS_WEBHOOK_URL: $(MS_TEAMS_WEBHOOK_URL)
+    CI: "true"
+    SYSTEM_ACCESSTOKEN: $(System.AccessToken)
+    CI_SFDX_HARDIS_AZURE_TOKEN: $(System.AccessToken)
+    SYSTEM_COLLECTIONURI: $(System.CollectionUri)
+    SYSTEM_TEAMPROJECT: $(System.TeamProject)
+    SYSTEM_JOB_DISPLAY_NAME: $(System.JobDisplayName)
+    SYSTEM_JOB_ID: $(System.JobId)
+    SYSTEM_PULLREQUEST_PULLREQUESTID: $(System.PullRequest.PullRequestId)
+    BUILD_REPOSITORY_ID: $(Build.Repository.ID)
+    BUILD_REPOSITORYNAME: $(Build.Repository.Name)
+    BUILD_SOURCEBRANCHNAME: $(Build.SourceBranchName)
+    BUILD_BUILD_ID: $(Build.BuildId)`
   }
 }